refactor(P11): split env.ts monolith into domain-grouped leaf modules
Behavior-preserving split of the remaining src/lib/env.ts monolith
(574 lines) into domain-grouped leaf modules, following the existing
env-startup.ts / env-validation.ts convention:

- env-getters-core.ts: Supabase URL/keys, root domain, site URL, demo
  URL, rate-limit backend, CRON_SECRET (+rotation ts), booking token
  secret, profile-header HMAC key, PHI/backup encryption keys, pooler URL
- env-getters-integrations.ts: Stripe, R2, Cloudflare API/DNS, Workers
  AI (sync + async, nosemgrep annotations preserved verbatim), Twilio,
  Meta/WhatsApp, email/SMTP, Anthropic, E2B, insurance
- env-getters-observability.ts: Plausible, GA, AV-scan
- env-flags.ts: isAiDisabledByEnv, isMfaEnabled, isSuperAdminMfaRequired,
  isCustomDomainsEnabled, isAdminGeoRestrictionEnabled,
  getGeoRestrictAdminCountries, isProduction, isCi, getWorkerEnv,
  getAllowStagingDestructiveCrons

src/lib/env.ts is now a thin barrel (251 lines) that keeps
enforceEnvValidation() (unchanged) and re-exports the identical
65-export public API. Verified with the TS compiler that the export
surface is byte-for-byte identical; no consumer import changes.

Guard updates required by the file moves (policy intent unchanged):
- .semgrep/env-access.yml: allowlist the env-* leaf modules (including
  the previously-extracted env-validation.ts / env-startup.ts, which
  were missed when they were split out)
- ci.yml R-08 CRON_SECRET guard: allow env-getters-core.ts

Docs: deep_dive_analysis.md P11 marked resolved (status table,
section 4, mermaid diagram, priority table).

Verification: tsc --noEmit 0 errors, eslint clean on touched files,
knip ratchet at baseline (0 unused files / 141 unused exports),
vitest 181 files / 2041 tests passed, prettier clean, R-08 grep clean.

Resolves P11.
