+++ alayse and acrchiculture/deep_dive_analysis.md
@@ -20,7 +20,7 @@ The items below were addressed in a follow-up pass.
 | P5 — `super-admin-actions.ts` god file                   | ✅ Major structural reduction completed          | `src/lib/super-admin-actions.ts` is now a thin public wrapper surface. Implementation bodies are split across 16 files in `src/lib/super-admin/`: `activity-helpers.ts`, `base.ts`, `billing-actions.ts`, `catalog-helpers.ts`, `clinic-detail-actions.ts`, `clinic-lifecycle-actions.ts`, `clinic-setup-actions.ts`, `dashboard-actions.ts`, `feature-actions.ts`, `helpers.ts`, `models.ts`, `promotion-helpers.ts`, `promotions-actions.ts`, `staff-provisioning-actions.ts`, `subscription-helpers.ts`, `types.ts`. The former `clinic-actions.ts` was split into the three `clinic-*-actions.ts` modules; `provisioning-actions.ts` became `staff-provisioning-actions.ts`.                                                                                                                                                                                                                                                                                  |
 | P8 — `withAuth` duplication (SEALED)                     | ⏸ Not attempted                                  | Left as-is: high effort/risk, and P8 specifically touches a SEALED file (`with-auth.ts`) that TASK-ROUTER says not to modify without explicit request.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 |
 | P9 — product boundary                                    | ⏸ Not attempted                                  | Left as-is. Recommend tackling this as its own dedicated, reviewed change rather than bundling it into a general cleanup pass.                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         |
-| P11 — `env.ts` monolith                                  | 🟡 Partially addressed                           | Extracted the production startup hard-fail guards and security-posture flag policy list into `src/lib/env-startup.ts`, then extracted the env-rule registry and `validateEnv()` into `src/lib/env-validation.ts`. `src/lib/env.ts` dropped from **1,498** lines to **574** lines.                                                                                                                                                                                                                                                                                                                                                                                                                                                      |
+| P11 — `env.ts` monolith                                  | ✅ Resolved                                      | Extracted the production startup hard-fail guards and security-posture flag policy list into `src/lib/env-startup.ts`, then extracted the env-rule registry and `validateEnv()` into `src/lib/env-validation.ts`, then split the remaining ~50 typed getters and toggle helpers into domain-grouped leaf modules: `src/lib/env-getters-core.ts` (Supabase, domains, platform secrets), `src/lib/env-getters-integrations.ts` (Stripe, R2, Cloudflare, Workers AI, Twilio, Meta/WhatsApp, email, Anthropic, E2B, insurance), `src/lib/env-getters-observability.ts` (Plausible, GA, AV-scan), and `src/lib/env-flags.ts` (boolean toggles + runtime-mode helpers). `src/lib/env.ts` is now a thin barrel (**251** lines, down from **1,498**) that keeps `enforceEnvValidation()` and re-exports the identical 65-export public API — no consumer import changed.                                                                                                                                                                                                                                                                                                                                                                                                                                                      |
 
 Validation for the fixed items: targeted pricing/revenue/tenant suites pass, `src/lib/__tests__/delete-clinic.test.ts` passes after the final wrapper cleanup, and a fresh full `npx vitest run --maxWorkers=1` still reports **179 test files passed**, **1996 tests passed**, **84 skipped**. Targeted ESLint on every touched super-admin file is clean.
 
@@ -100,9 +100,9 @@ The migration narrative (`00187_drop_clinical_emr_surface.sql`) claims Oltigo is
 `src/lib/with-auth.ts` contains two long wrappers that repeat nearly the entire flow (Supabase client creation, auth lookup, signed profile-header verification, tenant mismatch assertion, etc.).
 The substantive behavioral difference is role enforcement. Everything else is duplicated and must be fixed twice. _(Note: This file is SEALED per TASK-ROUTER)._
 
-### 4. 🟡 `env.ts` Is Still A Monolith (P11)
+### 4. ✅ `env.ts` Monolith Resolved (P11)
 
-While partially improved (dropped from 1,498 to 574 lines), it still centralizes environment parsing, grouped validation rules, getter helpers, production-mode helpers, and AI provider resolution. This makes change review and dependency hygiene harder than necessary.
+Resolved. `src/lib/env.ts` is now a thin barrel (251 lines) over six domain-grouped leaf modules — `env-validation.ts` (ENV_RULES + `validateEnv()`), `env-startup.ts` (production enforcement guards), `env-getters-core.ts` (Supabase, domains, platform secrets), `env-getters-integrations.ts` (Stripe, R2, Cloudflare, Workers AI, Twilio, Meta/WhatsApp, email, Anthropic, E2B, insurance), `env-getters-observability.ts` (Plausible, GA, AV-scan), and `env-flags.ts` (boolean toggles + runtime-mode helpers). The barrel keeps `enforceEnvValidation()` and preserves the exact 65-export public API, so `@/lib/env` remains the only import path consumers use. The `.semgrep/env-access.yml` allowlist and the R-08 CRON_SECRET CI guard were extended to cover the leaf modules.
 
 ---
 
@@ -154,7 +154,7 @@ graph TD
     P3["🟡 P3: Identity vocabulary unified<br/>capabilities.ts (enforcement = follow-up)"]
     P8["🟡 P8: Deduplicate withAuth wrappers (SEALED)"]
     P9["🟡 P9: Rationalize product boundary<br/>Clinical vs operational vs adjacent verticals"]
-    P11["🟡 P11: Split env.ts"]
+    P11["✅ P11: env.ts split into<br/>domain-grouped leaf modules"]
 
     P3 --> P8
     P9 --> P11
@@ -165,4 +165,4 @@ graph TD
 | 🟡 P3    | Role/slugs/persona identity drift | Vocabulary/prefix drift unified in `capabilities.ts`; capability-based enforcement is a follow-up (SEALED middleware) | Low | Partial |
 | 🟡 P8    | `withAuth` duplication (SEALED)   | Maintenance risk        | Medium | Low    |
 | 🟡 P9    | Product-boundary sprawl           | Strategic inconsistency | Medium | High   |
-| 🟡 P11   | `env.ts` monolith                 | Maintainability drag    | Low    | Medium |
+| ✅ P11   | `env.ts` monolith                 | Resolved — barrel + domain-grouped leaf modules, public API unchanged | Low | Done |
